refactor: optimize Search widget hooks
