+++ src/pages/Home.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Search } from '../widgets/Search';
+import { Search } from '../widgets/Search/Search';
 import { TestError } from '../entities/TestError';
 import { ErrorBoundary } from './ErrorBoundary';
 import { Outlet } from 'react-router-dom';
