+++ src/shared/hooks.ts
@@ -1,6 +1,5 @@
-import { SetStateAction, useState } from 'react';
+import { SetStateAction } from 'react';
 import { useSearchParams } from 'react-router-dom';
-import { getSearchTermFromLS, setSearchTermToLS } from '../utils/localStorage';
 
 export function usePage(): [number, (s: SetStateAction<number>) => void] {
   const [searchParams, setSearchParams] = useSearchParams();
@@ -14,27 +13,4 @@ export function usePage(): [number, (s: SetStateAction<number>) => void] {
   };
 
   return [page, setPage];
-}
-
-export function useSearchTerm(): [string, (s: string) => void] {
-  const [, setPage] = usePage();
-  const [searchTerm, setSearchTerm] = useState<string>(getSearchTermFromLS());
-  const updateSearchTerm = (searchTerm: string) => {
-    setSearchTerm(searchTerm);
-    setSearchTermToLS(searchTerm);
-    setPage(1);
-  };
-  return [searchTerm, updateSearchTerm];
-}
-
-export function useLimit(
-  defaultLimit: number
-): [number, (s: SetStateAction<number>) => void] {
-  const [, setPage] = usePage();
-  const [limit, setLimit] = useState(defaultLimit);
-  const setLimitResetPage = (limit: SetStateAction<number>) => {
-    setLimit(limit);
-    setPage(1);
-  };
-  return [limit, setLimitResetPage];
 }
